requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-hCaptcha==0.5.4
flatbuffers==1.12
google-api-core==1.28.0
google-auth==1.30.1
google-cloud-core==1.6.0
google-cloud-logging==2.4.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
